added brightness adjustment to Modal Lights
