+++ .vscode/settings.json
@@ -46,5 +46,6 @@
     "streambuf": "cpp",
     "cinttypes": "cpp",
     "typeinfo": "cpp"
-  }
+  },
+  "cmake.configureOnOpen": false
 }
