+++ lib/ModalLights/include/lightDefines.h
@@ -1,8 +1,12 @@
 /*
- * This file creates two complimentary square waves that
- * create an ac signal, intended to drive LEDs wired in
- * opposite directions. Changing the brightness will
- * change both LEDs equally.
+ * This file contains defines necessary for the Modal controller
+ *
+ * It includes defines relating to the hardware PWM
+ * (i.e. maximum duty cycle). ModalLights should retain ownership
+ * of these defines, because they relate directly to the library
+ * and the code that interfaces with the library, but also
+ * because this is the only piece of code that should interact
+ * with the hardware abstraction layer.
  */
 
 #ifndef __LIGHT_DEFINES_H__
@@ -32,9 +36,19 @@ struct LightStateStruct {
   /**
    * @brief returns the brightness of the lights
   */
-  duty_t brightness(){
+  duty_t getBrightness(){
     return dutyLevel * state;
   };
+
+  /**
+   * @brief Sets the dutyLevel and state. doesn't check limits
+   * 
+   * @param brightness 
+   */
+  void setBrightness(duty_t brightness){
+    dutyLevel = brightness;
+    state = brightness > 0;
+  }
 };
 
 #endif
