--- lib/ModalLights/src/ModalLights.cpp
@@ -1,5 +1,0 @@
-#include "ModalLights.h"
-
-#ifdef native_env
-  #include <iostream>
-#endif
