+++ platformio.ini
@@ -40,7 +40,7 @@ build_flags =
 	-std=c++14
 debug_build_flags = 
 	-std=c++14
-debug_test = test_ModalLights
+; debug_test = test_ModalLights
 build_type = debug
 ; lib_deps =
 ; 	RTC_interface=symlink://{$workspace}/lib/RTC_interface/include
